# Only compute Position if not already in state and optimize line calculation