--- a/packages/babel-parser/src/parser/location.js
+++ b/packages/babel-parser/src/parser/location.js
@@ -13,8 +13,8 @@ export default class LocationParser extends CommentsParser {
   getLocationForPosition(pos: number): Position {
     let loc;
     if (pos === this.state.start) loc = this.state.startLoc;
-    else if (pos === this.state.end) loc = this.state.endLoc;
     else if (pos === this.state.lastTokStart) loc = this.state.lastTokStartLoc;
+    else if (pos === this.state.end) loc = this.state.endLoc;
     else if (pos === this.state.lastTokEnd) loc = this.state.lastTokEndLoc;
     else loc = getLineInfo(this.input, pos);