| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | 
| Patch: Bug Fix?          | n
| Major: Breaking Change?  | n
| Minor: New Feature?      | n
| Tests Added + Pass?      | Yes
| Documentation PR Link    | <!-- If only readme change, add `[skip ci]` to your commits -->
| Any Dependency Changes?  |
| License                  | MIT

Because of #9974 I was curious why throwing an exception is slow. So I did some investigation and my first guess was maybe we do something really slow when throwing. Turns out we do. 😫Every time we throw an exception we do calculate the line and column based on the current position (index) in the string.

So I was curious if that can be improved and after looking through StackOverflow for a long time :P I couldn't get this algorithm any faster.



After thinking more about I realized that actually most errors are thrown right when encountered, means the state is currently pointing to the error position. In this case and even in the case when the error is reported for the last token, we can easily reuse the location that is already in the state. No need for calculation at all.

> All benchmarks are node 12.2.0

<details><summary>Before</summary>

```
┌──────────────────────────┬───────────────────────────────┬─────────────────────────────┬─────────────────────────────┐
│ fixture                  │ babel                         │ babelWithFlow               │ babelWithTS                 │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/angular.js           │ 26.54 ops/sec ±4.28% (38ms)   │ 11.1 ops/sec ±3.25% (90ms)  │ 4.76 ops/sec ±2.89% (210ms) │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/ember.debug.js       │ 10.56 ops/sec ±2.77% (95ms)   │ 4.4 ops/sec ±2.91% (227ms)  │ 1.04 ops/sec ±9.82% (964ms) │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/babylon-dist.js      │ 44.52 ops/sec ±5.35% (22ms)   │ 18.07 ops/sec ±2.87% (55ms) │ 16.17 ops/sec ±2.41% (62ms) │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/jquery.js            │ 52.38 ops/sec ±5.04% (19ms)   │ 20.95 ops/sec ±4.67% (48ms) │ 16.15 ops/sec ±3.33% (62ms) │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/backbone.js          │ 229 ops/sec ±24.19% (4.359ms) │ 90.2 ops/sec ±13.65% (11ms) │ 90.85 ops/sec ±2.16% (11ms) │
├──────────────────────────┼───────────────────────────────┼─────────────────────────────┼─────────────────────────────┤
│ es5/react-with-addons.js │ 24.81 ops/sec ±2.63% (40ms)   │ 10.44 ops/sec ±3.12% (96ms) │ 7.17 ops/sec ±2.35% (140ms) │
└──────────────────────────┴───────────────────────────────┴─────────────────────────────┴─────────────────────────────┘
```

</details>

<details><summary>After</summary>

```
┌──────────────────────────┬───────────────────────────────┬──────────────────────────────┬──────────────────────────────┐
│ fixture                  │ babel                         │ babelWithFlow                │ babelWithTS                  │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/angular.js           │ 27.69 ops/sec ±4.56% (36ms)   │ 11.52 ops/sec ±3.56% (87ms)  │ 12.58 ops/sec ±2.87% (79ms)  │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/ember.debug.js       │ 11.07 ops/sec ±3.31% (90ms)   │ 4.48 ops/sec ±3.77% (223ms)  │ 4.95 ops/sec ±2.51% (202ms)  │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/babylon-dist.js      │ 46.94 ops/sec ±1.58% (21ms)   │ 18.49 ops/sec ±2.76% (54ms)  │ 20 ops/sec ±4.82% (50ms)     │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/jquery.js            │ 56.61 ops/sec ±1.66% (18ms)   │ 21.94 ops/sec ±4.18% (46ms)  │ 23.57 ops/sec ±4.7% (42ms)   │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/backbone.js          │ 230 ops/sec ±28.92% (4.351ms) │ 95.37 ops/sec ±13.28% (10ms) │ 108 ops/sec ±1.79% (9.263ms) │
├──────────────────────────┼───────────────────────────────┼──────────────────────────────┼──────────────────────────────┤
│ es5/react-with-addons.js │ 26.29 ops/sec ±3.04% (38ms)   │ 10.98 ops/sec ±3.16% (91ms)  │ 12.07 ops/sec ±2.86% (83ms)  │
└──────────────────────────┴───────────────────────────────┴──────────────────────────────┴──────────────────────────────┘
```

</details>

So together with #9974 this should give a good boost for ts in certain situations

@matthewrobertson Can you check if that improves your reported case even further?