@matthewrobertson 2019-05-16
Nice find @danez! I think with this change we might not need #9974. I did some simple profiling and it looks like with this change alone we get most of the perf benefits that would be gained from that PR:

parser config | parse time
------------ | -------------
no typescript | 2.25s
with typescript | 8.05s
with typescript AND #9974: | 3.33s
with typescript AND #9989: | 3.18s
with typescript AND #9989 AND #9974 | 2.59s

So there is some additional benefit from adding #9974 to this, but this change alone is better than #9974 alone. Maybe I can trim that PR down a bit so we just take some of the simpler optimizations that don't add too much complexity.