@danez 2019-05-16
> Maybe I can trim that PR down a bit so we just take some of the simpler optimizations that don't add too much complexity.

I think your PR is still very valuable and I don't see why we shouldn't bail out as soon as possible rather than try parsing and reverting state etc.